Moves subdomain urls and reverse
