+++ app/grandchallenge/datasets/views.py
@@ -4,7 +4,7 @@ from django.views.generic import ListView, CreateView, DetailView, UpdateView
 from grandchallenge.cases.forms import UploadRawImagesForm
 from grandchallenge.cases.models import RawImageUploadSession
 from grandchallenge.core.permissions.mixins import UserIsStaffMixin
-from grandchallenge.subdomains.urls import reverse
+from grandchallenge.subdomains.utils import reverse
 from grandchallenge.datasets.forms import (
     ImageSetUpdateForm,
     AnnotationSetForm,
