+++ app/grandchallenge/evaluation/emails.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.core.mail import send_mail
 
-from grandchallenge.subdomains.urls import reverse
+from grandchallenge.subdomains.utils import reverse
 from grandchallenge.evaluation.templatetags.evaluation_extras import user_error
