+++ app/grandchallenge/evaluation/models.py
@@ -13,7 +13,7 @@ from grandchallenge.container_exec.models import (
     ContainerImageModel,
 )
 from grandchallenge.core.models import UUIDModel
-from grandchallenge.subdomains.urls import reverse
+from grandchallenge.subdomains.utils import reverse
 from grandchallenge.core.validators import (
     MimeTypeValidator,
     ExtensionValidator,
