+++ app/grandchallenge/evaluation/views.py
@@ -11,7 +11,7 @@ from grandchallenge.core.permissions.mixins import (
     UserIsChallengeAdminMixin,
     UserIsChallengeParticipantOrAdminMixin,
 )
-from grandchallenge.subdomains.urls import reverse
+from grandchallenge.subdomains.utils import reverse
 from grandchallenge.evaluation.forms import (
     MethodForm,
     SubmissionForm,
