+++ app/grandchallenge/profiles/views.py
@@ -3,7 +3,7 @@ from django.shortcuts import redirect, render
 from userena import views as userena_views
 
 from grandchallenge.profiles.forms import EditProfileForm
-from grandchallenge.subdomains.urls import reverse
+from grandchallenge.subdomains.utils import reverse
 
 
 def login_redirect(request):
