+++ app/grandchallenge/uploads/views.py
@@ -10,7 +10,7 @@ from grandchallenge.core.permissions.mixins import UserIsChallengeAdminMixin
 from grandchallenge.core.views import getSite, permissionMessage
 from grandchallenge.pages.models import Page
 from grandchallenge.pages.views import ChallengeFilteredQuerysetMixin
-from grandchallenge.subdomains.urls import reverse
+from grandchallenge.subdomains.utils import reverse
 from grandchallenge.uploads.emails import send_file_uploaded_notification_email
 from grandchallenge.uploads.forms import UserUploadForm
 from grandchallenge.uploads.models import UploadModel
