+++ app/tests/common_tests/test_link_visibility.py
@@ -1,6 +1,6 @@
 import pytest
 
-from grandchallenge.subdomains.urls import reverse
+from grandchallenge.subdomains.utils import reverse
 from tests.utils import validate_admin_only_text_in_page
